requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

scrapy>=2.11.0
requests>=2.31.0
aiohttp>=3.9.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
readability-lxml>=0.8.1
markdownify>=0.13.1

cloudscraper>=1.2.71
selenium>=4.15.0
undetected-chromedriver>=3.5.5
playwright>=1.52.0

sqlalchemy>=2.0.0
aiosqlite>=0.19.0
greenlet>=3.0.0

pyyaml>=6.0
toml>=0.10.0

python-dateutil>=2.8.0
pytz>=2023.3
tqdm>=4.66.0
httpx>=0.27.0
python-dotenv>=1.0.0
curl_cffi>=0.7.0
openpyxl>=3.1.0
mammoth>=1.8.0
PyMuPDF>=1.24.0
Pillow>=10.0.0

loguru>=0.7.0
prometheus-client>=0.19.0

pytest>=7.4.0
pytest-asyncio>=0.21.0
